Move _repr_header.
