+++ dolfyn/__init__.py
@@ -1,3 +1,4 @@
 from ._version import __version__
 from .main import *
 from .data.velocity import VelBinner
+from .rotate import rotate
